requirements.txt
streamlit
supabase
fpdf2
pandas